refactored, added interthread messaging.

+++ src/Thread.js
@@ -5,10 +5,10 @@ class Thread {
 		this._worker = null;
 		this._threadBody = null;
 		this._threadParamNames = null;
-		this._completedCallbackName = null;
 		this._parseUserFunction(func);
 		this.running = false;
 		this.thread_id = Thread.instances.length;
+		this.onmessage = ()=>{};
 		Thread.instances.push(this);
 	}
 
@@ -18,19 +18,41 @@ class Thread {
 		}
 		var fstr = func.toString();
 		this._threadBody = fstr.substring(fstr.indexOf("{") + 1, fstr.lastIndexOf("}"));
-		this._threadParamNames = fstr.substring(fstr.indexOf("(") + 1, fstr.indexOf(")")).split(",").map(p => p.trim());
-		this._completedCallbackName = this._threadParamNames.pop();
+		this._threadParamNames = fstr.substring(fstr.indexOf("(") + 1, fstr.indexOf(")")).split(",").map(p => p.trim()).filter(p=>!!p);
 	}
 
 	_generateWorker(){
-		var workerStr = `var ${this._completedCallbackName} = function(){
-			var args = Array.from(arguments);
-			postMessage(args);
-		};
+		var workerStr = `class ThreadMain{
+			constructor(){
+				this.running = false;
+				this.onmessage = () => {};
+			}
+
+			run(...params){
+				this.running = true;
+				var [${this._threadParamNames.join(", ")}] = params;
+				${this._threadBody}
+			}
+
+			done(...params){
+				self.postMessage({isdone: true, params: [...params]});
+				this.running = false;
+			}
+
+			postmessage(...params){
+				self.postMessage({isdone: false, params: [...params]});
+			}
+		}
+
+		var thread = new ThreadMain();
+
 		self.onmessage = function(d){
-			var [${this._threadParamNames.join(", ")}] = d.data;
-			${this._threadBody}
-		};`;
+			if(!thread.running){
+				thread.run(...d.data);
+			}else{
+				thread.onmessage(...d.data);
+			}
+		};`;		
 		var blob = new Blob([workerStr], {type: 'application/javascript'});
 		this._worker = new Worker(URL.createObjectURL(blob));
 	}
@@ -51,12 +73,21 @@ class Thread {
 				
 			};
 			this._worker.onmessage = d => {
-				this._worker.terminate();
-				resolve(...d.data);
+				if(d.data.isdone){
+					this._worker.terminate();
+					resolve(...d.data.params);
+				}else{
+					this.onmessage(...d.data.params);
+				}
 			};
 			this._worker.postMessage(this._params);
 			this.running = true;
 		});
+		return this;
+	}
+	
+	postmessage(...params){
+		this._worker.postMessage(params);
 		return this;
 	}
